BUG: Use size_t to compute CUDA memory length
Fixes #594. Opened files with command line
$(find -type f -exec grep -lE 'cuda(Malloc|Memset|Memcpy)' {} \;)
and searched for occurences with vim regular expression
cuda\(Malloc\|Memset\|Memcpy\)

+++ src/rtkCudaConjugateGradientImageFilter.cu
@@ -31,7 +31,7 @@ void
 CUDA_copy(long int numberOfElements, float * in, float * out)
 {
   // Copy input volume to output
-  long int memorySizeOutput = numberOfElements * sizeof(float);
+  size_t memorySizeOutput = numberOfElements * sizeof(float);
   cudaMemcpy(out, in, memorySizeOutput, cudaMemcpyDeviceToDevice);
 }
 
@@ -39,7 +39,7 @@ void
 CUDA_copy(long int numberOfElements, double * in, double * out)
 {
   // Copy input volume to output
-  long int memorySizeOutput = numberOfElements * sizeof(double);
+  size_t memorySizeOutput = numberOfElements * sizeof(double);
   cudaMemcpy(out, in, memorySizeOutput, cudaMemcpyDeviceToDevice);
 }
